refactor: 틀린 클래스명 정정
